% Generated by roxygen2: do not edit by hand
% Please edit documentation in r/whisk.R
\name{whisk}
\alias{whisk}
\title{Tidy the result of a test into a nested summary data.frame.}
\usage{
whisk(x)
}
\arguments{
\item{x}{an object to be passed to \code{broom::tidy} + \code{broom::glance}}
}
\value{
A data.frame
}
\description{
Combines broom::tidy, broom::glance, and tidyr::nest into a single function
that takes any model compatible with broom and returns a one-row data frame
with the nested model and data from broom::glance.
}
